+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add Python 3.12 support
+- Add colour-coded row background based on cert expiry date #1
 
 ### Changed
 - Use Poetry to manage project